fixed WSOD
