+++ .storybook/main.js
@@ -2,7 +2,7 @@ module.exports = {
   stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     '@storybook/addon-links',
-    '@storybook/addon-essentials',
+    // '@storybook/addon-essentials',
     '@storybook/preset-create-react-app',
   ],
   webpackFinal: async (config) => {
@@ -13,7 +13,7 @@ module.exports = {
         alias: {
           ...config.resolve.alias,
           '@emotion/core': '@emotion/react',
-          'emotion-theming': '@emotion/react'
+          'emotion-theming': '@emotion/react',
         },
       },
     };
